Move coldrewards/budget code into devault directory, rm mnemonic directory to consolidate files
