+++ .gitignore
@@ -1,7 +1,6 @@
 *.tar.gz
 
 *.exe
-src/devault
 src/devaultd
 src/devault-cli
 src/devault-seeder
