+++ src/devault/CMakeLists.txt
@@ -3,28 +3,30 @@
 # Distributed under the MIT software license, see the accompanying
 # file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
-project(rewards)
+project(devault)
 
-file(GLOB REWARD_HEADERS ${CMAKE_CURRENT_SOURCE_DIR}/*.h)
-source_group("RewardHeaders" FILES ${REWARD_HEADERS})
+file(GLOB DEVAULT_HEADERS ${CMAKE_CURRENT_SOURCE_DIR}/*.h)
+source_group("RewardHeaders" FILES ${DEVAULT_HEADERS})
 
 
 # The library
-add_library(coldrewards
-  ${REWARD_HEADERS}
+add_library(devault
+  ${DEVUALT_HEADERS}
 	rewards.cpp
 	rewards_calculation.cpp
   rewardsview.cpp
+  budget.cpp
 )
 
 find_package(Boost 1.58 REQUIRED ${BOOST_PACKAGES_REQUIRED})
 
-target_include_directories(coldrewards
+target_include_directories(devault
 	PRIVATE
-		${CMAKE_CURRENT_SOURCE_DIR}/../leveldb/include
     ${Boost_INCLUDE_DIRS}
+		${CMAKE_CURRENT_SOURCE_DIR}/../leveldb/include
 		..
 	PUBLIC
 		# To access the config.
 		${CMAKE_CURRENT_BINARY_DIR}/..
 )
+
