+++ src/devault/budget.cpp
@@ -2,7 +2,7 @@
 // Copyright (c) 2019 Jon Spock
 // Distributed under the MIT software license, see the accompanying
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
-#include "budget/budget.h"
+#include "budget.h"
 #include "amount.h"
 #include "chainparams.h"
 #include "config.h"
