+++ src/devault/rewards_calculation.h
