+++ src/devault/rewardsview.h
