--- src/mnemonic/.clang-format
@@ -1,18 +1,0 @@
-Language:             Cpp
-Standard:             Cpp11
-ColumnLimit:          120
-IndentWidth:          2
-TabWidth:             4
-AccessModifierOffset: -4
-UseTab:               Never
-IndentCaseLabels:     true
-NamespaceIndentation: None
-BreakBeforeBraces:    Attach
-ReflowComments:       true
-AlignEscapedNewlines: Right
-AllowShortBlocksOnASingleLine: true
-AllowShortIfStatementsOnASingleLine: true
-AllowShortFunctionsOnASingleLine: All
-FixNamespaceComments: false
-ForEachMacros:   [ foreach, Q_FOREACH, BOOST_FOREACH, BOOST_REVERSE_FOREACH ]
-CommentPragmas:  '@DISABLE FORMATING FOR THIS COMMENT@'
