+++ src/validation.cpp
@@ -44,8 +44,8 @@
 #include "validationinterface.h"
 #include "warnings.h"
 
-#include "coldrewards/rewards.h"
-#include "budget/budget.h"
+#include "devault/rewards.h"
+#include "devault/budget.h"
 
 #include <atomic>
 #include <future>
